fix: :bug: fix null comparison in 07/2020 Nikofix
Fixed a bug in the prior 07/2020 Nikofix that would cause processing of some single image post to stop before adding it to downloads. This fix should be non-intrusive and pick up formerly ignored posts to add them to the download queue. Also added version numbers alongside dates in the options page news section for easier classification of news.