feat: :sparkles: add download interval setting
As per user request in issue #3, a setting was introduced that lets the user manipulate the download schedule interval. The setting handling has been reworked (especially the part below `browser.storage.local.get('settings').then((result => {`, which should not have worked!) and the download interval initialization has been rewritten to enable cancelation and restart of it whenever the interval changes.